# [NER] improve inference perf by > 40% on GPU when CRF is used