--- a/flair/models/sequence_tagger_model.py
+++ b/flair/models/sequence_tagger_model.py
@@ -170,15 +170,15 @@ class SequenceTagger(flair.nn.Model):
                 self.embeddings.embedding_length, len(tag_dictionary)
             )
 
-        self.crf_transitions = torch.nn.Parameter(
+        self.transitions = torch.nn.Parameter(
             torch.randn(self.tagset_size, self.tagset_size)
         )
 
-        self.crf_transitions.detach()[
+        self.transitions.detach()[
             self.tag_dictionary.get_idx_for_item(START_TAG), :
         ] = -10000
 
-        self.crf_transitions.detach()[
+        self.transitions.detach()[
             :, self.tag_dictionary.get_idx_for_item(STOP_TAG)
         ] = -10000
 
@@ -248,7 +248,7 @@ class SequenceTagger(flair.nn.Model):
             metric = Metric("Evaluation")
 
             lines: List[str] = []
-            crf_transitions = self.crf_transitions.detach().cpu()
+            transitions = self.transitions.detach().cpu()
 
             for batch in data_loader:
                 batch_no += 1
@@ -256,7 +256,7 @@ class SequenceTagger(flair.nn.Model):
                 with torch.no_grad():
                     features = self.forward(batch)
                     loss = self._calculate_loss(features, batch)
-                    tags, _ = self._obtain_labels(features, batch, crf_transitions)
+                    tags, _ = self._obtain_labels(features, batch, transitions)
 
                 eval_loss += loss
 
@@ -353,7 +353,7 @@ class SequenceTagger(flair.nn.Model):
             # reverse sort all sequences by their length
             filtered_sentences.sort(key=lambda x: len(x), reverse=True)
 
-            crf_transitions = self.crf_transitions.detach().cpu()
+            transitions = self.transitions.detach().cpu()
 
             # make mini-batches
             batches = [
@@ -372,7 +372,7 @@ class SequenceTagger(flair.nn.Model):
 
                 feature = self.forward(batch)
                 tags, all_tags = self._obtain_labels(
-                    feature, batch, crf_transitions, get_all_tags=all_tag_prob
+                    feature, batch, transitions, get_all_tags=all_tag_prob
                 )
 
                 for (sentence, sent_tags) in zip(batch, tags):
@@ -490,7 +490,7 @@ class SequenceTagger(flair.nn.Model):
             r = torch.LongTensor(range(lens_[i])).to(flair.device)
 
             score[i] = torch.sum(
-                self.crf_transitions[
+                self.transitions[
                     pad_stop_tags[i, : lens_[i] + 1], pad_start_tags[i, : lens_[i] + 1]
                 ]
             ) + torch.sum(feats[i, r, tags[i, : lens_[i]]])
@@ -542,7 +542,7 @@ class SequenceTagger(flair.nn.Model):
         self,
         feature: torch.Tensor,
         sentences: List[Sentence],
-        crf_transitions: Parameter,
+        transitions: Parameter,
         get_all_tags: bool = False,
     ) -> (List[List[Label]], List[List[List[Label]]]):
         """
@@ -561,9 +561,7 @@ class SequenceTagger(flair.nn.Model):
         for feats, length in zip(feature, lengths):
             if self.use_crf:
                 confidences, tag_seq, scores = self._viterbi_decode(
-                    feats[:length],
-                    all_scores=get_all_tags,
-                    crf_transitions=crf_transitions,
+                    feats[:length], all_scores=get_all_tags, transitions=transitions
                 )
             else:
                 tag_seq = []
@@ -599,7 +597,7 @@ class SequenceTagger(flair.nn.Model):
 
         return tags, all_tags
 
-    def _viterbi_decode(self, feats, crf_transitions, all_scores: bool = False):
+    def _viterbi_decode(self, feats, transitions, all_scores: bool = False):
         backpointers = []
         backscores = []
 
@@ -610,7 +608,7 @@ class SequenceTagger(flair.nn.Model):
         for feat in feats:
             next_tag_var = (
                 forward_var.view(1, -1).expand(self.tagset_size, self.tagset_size)
-                + crf_transitions
+                + transitions
             )
             _, bptrs_t = torch.max(next_tag_var, dim=1)
             viterbivars_t = next_tag_var[range(len(bptrs_t)), bptrs_t]
@@ -619,8 +617,7 @@ class SequenceTagger(flair.nn.Model):
             backpointers.append(bptrs_t)
 
         terminal_var = (
-            forward_var
-            + crf_transitions[self.tag_dictionary.get_idx_for_item(STOP_TAG)]
+            forward_var + transitions[self.tag_dictionary.get_idx_for_item(STOP_TAG)]
         )
         terminal_var.detach()[self.tag_dictionary.get_idx_for_item(STOP_TAG)] = -10000.0
         terminal_var.detach()[
@@ -683,25 +680,25 @@ class SequenceTagger(flair.nn.Model):
 
         forward_var[:, 0, :] = init_alphas[None, :].repeat(feats.shape[0], 1)
 
-        crf_transitions = self.crf_transitions.view(
-            1, self.crf_transitions.shape[0], self.crf_transitions.shape[1]
+        transitions = self.transitions.view(
+            1, self.transitions.shape[0], self.transitions.shape[1]
         ).repeat(feats.shape[0], 1, 1)
 
         for i in range(feats.shape[1]):
             emit_score = feats[:, i, :]
 
             tag_var = (
-                emit_score[:, :, None].repeat(1, 1, crf_transitions.shape[2])
-                + crf_transitions
+                emit_score[:, :, None].repeat(1, 1, transitions.shape[2])
+                + transitions
                 + forward_var[:, i, :][:, :, None]
-                .repeat(1, 1, crf_transitions.shape[2])
+                .repeat(1, 1, transitions.shape[2])
                 .transpose(2, 1)
             )
 
             max_tag_var, _ = torch.max(tag_var, dim=2)
 
             tag_var = tag_var - max_tag_var[:, :, None].repeat(
-                1, 1, crf_transitions.shape[2]
+                1, 1, transitions.shape[2]
             )
 
             agg_ = torch.log(torch.sum(torch.exp(tag_var), dim=2))
@@ -713,7 +710,7 @@ class SequenceTagger(flair.nn.Model):
 
         forward_var = forward_var[range(forward_var.shape[0]), lens_, :]
 
-        terminal_var = forward_var + self.crf_transitions[
+        terminal_var = forward_var + self.transitions[
             self.tag_dictionary.get_idx_for_item(STOP_TAG)
         ][None, :].repeat(forward_var.shape[0], 1)
 
@@ -884,7 +881,7 @@ class SequenceTagger(flair.nn.Model):
 
     def get_transition_matrix(self):
         data = []
-        for to_idx, row in enumerate(self.crf_transitions):
+        for to_idx, row in enumerate(self.transitions):
             for from_idx, column in enumerate(row):
                 row = [
                     self.tag_dictionary.get_item_for_index(from_idx),