Viterbi seems to be the bottle neck of the inference part when GPU is used with CRF.

Here we simply move the computation of this part from GPU to CPU as the operations are not GPU optimized.
On my 2080TI with I7 (6 real cores), on the same data / same model, I went from 1mn 33s to 63 seconds inference time.
I have not made any test on other hardware but I expect perf improvement to be similar.

Moreover, this open the door to more optimizations as it removes the "lock" on GPU (we can easily do a better job at multithreading the process than Pytorch is currently doing), plus the same performance boost can probably be reproduced on the training part.