@pommedeterresautee 2019-08-27
I have made some tests to move the whole CRF part on the CPU (including for the training), keeping the featurisation on GPU, and it brings no improvement at all for the training part. 

I still think there are performance opportunities remaining as the GPU use is low and the inference is still not multithreaded at its best. 

One strategy I want to try now that the Viterbi part is on CPU is multithreading the process at the batch level. Right now the Viterbi algo is applied serially to each sentence, however the features are extracted by the GPU in batch. Before this PR there was a GPU lock, but now Viterbi is CPU only it should be doable to distribute the computation on each CPU core.