@alanakbik 2019-08-29
Hello @pommedeterresautee thanks a lot for this PR - we've unhappy with the performance of the CRF decoder for a while now so this really helps! I've run a test on CoNLL-03 predicting with the current 'ner' model on a v100 and prediction time decreases **from ~20 to ~15 seconds** on the evaluation dataset and **from ~95 to ~60 seconds** on all sentences in CoNLL-03. This is great! 

I'll run a few more tests on some other datasets to make sure that we get the same eval numbers and will merge into master!