@stefan-it 2019-08-29
Not directly related to this issue: but I've seen some F-Score mismatches between Flair output and the CoNLL script (maybe we can switch back to the conll script, or at least with a configurable parameter 🤔) But I'll collect some examples and open a new issue for that!