@alanakbik 2019-08-29
@stefan-it yes good point - Flair uses a slightly different logic to get from tokens to spans which should account for the difference. But this is something we've been meaning to add back in so it's good to put this back on the agenda.