@alanakbik 2019-08-29
@pommedeterresautee the following code is throwing an error: 

```python
tagger = SequenceTagger.load("frame-fast")
```

This likely affects pre-trained models that were trained without a CRF. I'll check into this a bit more.