@alanakbik 2019-08-29
I think the problem is that `self.transitions` is now always assumed to be there, but sometimes models are trained with `use_crf` set to `False`, which then throws an error. To reprodue: 

```python
tagger = SequenceTagger.load("frame-fast")
tagger.predict(Sentence("I love Berlin"))
```

@pommedeterresautee could you take a look?