@pommedeterresautee 2019-08-29
You were right about the error, self.transitions is the culprit, to remove the error it would require to retrain every model. The commit avoid such behaviour, now the matrix is created only when required.